+++ mc2/controllers/base/tests/test_views.py
@@ -1,4 +1,7 @@
+import json
 import os
+import uuid
+
 import pytest
 import responses
 from django.conf import settings
@@ -6,8 +9,7 @@ from django.test import RequestFactory
 from django.test.client import Client
 from django.core.urlresolvers import reverse
 from django.contrib.auth.models import User
-from django.db.models.signals import post_save
-from mc2.controllers.base.models import Controller, publish_to_websocket
+from mc2.controllers.base.models import Controller
 from mc2.controllers.base.tests.base import ControllerBaseTestCase
 from mc2.controllers.base.tests.utils import setup_responses_for_logdriver
 from mc2.controllers.base.views import AppEventSourceView
@@ -23,8 +25,6 @@ class ViewsTestCase(ControllerBaseTestCase):
         self.client = Client()
         self.client.login(username='testuser', password='test')
         self.user = User.objects.get(username='testuser')
-
-        post_save.disconnect(publish_to_websocket, sender=Controller)
 
     @responses.activate
     def test_create_new_controller(self):
@@ -43,6 +43,10 @@ class ViewsTestCase(ControllerBaseTestCase):
             'env-INITIAL_FORMS': 0,
             'env-MIN_NUM_FORMS': 0,
             'env-MAX_NUM_FORMS': 100,
+            'label-TOTAL_FORMS': 0,
+            'label-INITIAL_FORMS': 0,
+            'label-MIN_NUM_FORMS': 0,
+            'label-MAX_NUM_FORMS': 100,
         }
 
         response = self.client.post(reverse('base:add'), data)
@@ -59,7 +63,7 @@ class ViewsTestCase(ControllerBaseTestCase):
         self.assertTrue(controller.slug)
 
     @responses.activate
-    def test_create_new_controller_with_env(self):
+    def test_create_new_controller_with_label(self):
         existing_controller = self.mk_controller()
 
         self.client.login(username='testuser2', password='test')
@@ -71,12 +75,62 @@ class ViewsTestCase(ControllerBaseTestCase):
         data = {
             'name': 'Another test app',
             'marathon_cmd': 'ping2',
+            'env-TOTAL_FORMS': 0,
+            'env-INITIAL_FORMS': 0,
+            'env-MIN_NUM_FORMS': 0,
+            'env-MAX_NUM_FORMS': 100,
+            'label-0-name': 'A_TEST_KEY',
+            'label-0-value': 'the value',
+            'label-TOTAL_FORMS': 1,
+            'label-INITIAL_FORMS': 0,
+            'label-MIN_NUM_FORMS': 0,
+            'label-MAX_NUM_FORMS': 100,
+        }
+
+        response = self.client.post(reverse('base:add'), data)
+
+        self.assertEqual(response.status_code, 302)
+
+        controller = Controller.objects.exclude(
+            pk=existing_controller.pk).get()
+        self.assertEqual(controller.state, 'done')
+
+        self.assertEqual(controller.name, 'Another test app')
+        self.assertEqual(controller.marathon_cmd, 'ping2')
+        self.assertEqual(controller.organization.slug, 'foo-org')
+        self.assertEqual(controller.label_variables.count(), 1)
+        self.assertEqual(
+            controller.label_variables.all()[0].name, 'A_TEST_KEY')
+        self.assertEqual(
+            controller.label_variables.all()[0].value, 'the value')
+        self.assertTrue(controller.slug)
+
+    @responses.activate
+    def test_create_new_controller_with_env(self):
+        existing_controller = self.mk_controller()
+
+        self.client.login(username='testuser2', password='test')
+        self.client.get(
+            reverse('organizations:select-active', args=('foo-org',)))
+
+        self.mock_create_marathon_app()
+
+        webhook_token = str(uuid.uuid4())
+
+        data = {
+            'name': 'Another test app',
+            'marathon_cmd': 'ping2',
             'env-0-key': 'A_TEST_KEY',
             'env-0-value': 'the value',
             'env-TOTAL_FORMS': 1,
             'env-INITIAL_FORMS': 0,
             'env-MIN_NUM_FORMS': 0,
             'env-MAX_NUM_FORMS': 100,
+            'label-TOTAL_FORMS': 0,
+            'label-INITIAL_FORMS': 0,
+            'label-MIN_NUM_FORMS': 0,
+            'label-MAX_NUM_FORMS': 100,
+            'webhook_token': webhook_token,
         }
 
         response = self.client.post(reverse('base:add'), data)
@@ -93,6 +147,7 @@ class ViewsTestCase(ControllerBaseTestCase):
         self.assertEqual(controller.env_variables.count(), 1)
         self.assertEqual(controller.env_variables.all()[0].key, 'A_TEST_KEY')
         self.assertEqual(controller.env_variables.all()[0].value, 'the value')
+        self.assertEqual(str(controller.webhook_token), webhook_token)
         self.assertTrue(controller.slug)
 
     @responses.activate
@@ -105,6 +160,10 @@ class ViewsTestCase(ControllerBaseTestCase):
             'env-INITIAL_FORMS': 0,
             'env-MIN_NUM_FORMS': 0,
             'env-MAX_NUM_FORMS': 100,
+            'label-TOTAL_FORMS': 1,
+            'label-INITIAL_FORMS': 0,
+            'label-MIN_NUM_FORMS': 0,
+            'label-MAX_NUM_FORMS': 100,
         }
         response = self.client.post(reverse('base:add'), data)
 
@@ -118,6 +177,10 @@ class ViewsTestCase(ControllerBaseTestCase):
             'env-INITIAL_FORMS': 0,
             'env-MIN_NUM_FORMS': 0,
             'env-MAX_NUM_FORMS': 100,
+            'label-TOTAL_FORMS': 1,
+            'label-INITIAL_FORMS': 0,
+            'label-MIN_NUM_FORMS': 0,
+            'label-MAX_NUM_FORMS': 100,
         }
         response = self.client.post(reverse('base:add'), data)
 
@@ -169,6 +232,10 @@ class ViewsTestCase(ControllerBaseTestCase):
             'env-INITIAL_FORMS': 0,
             'env-MIN_NUM_FORMS': 0,
             'env-MAX_NUM_FORMS': 100,
+            'label-TOTAL_FORMS': 0,
+            'label-INITIAL_FORMS': 0,
+            'label-MIN_NUM_FORMS': 0,
+            'label-MAX_NUM_FORMS': 100,
         }
 
         response = self.client.post(reverse('base:add'), data)
@@ -176,6 +243,8 @@ class ViewsTestCase(ControllerBaseTestCase):
 
         controller = Controller.objects.all().last()
         self.mock_update_marathon_app(controller.app_id)
+
+        webhook_token = str(uuid.uuid4())
 
         self.client.post(
             reverse('base:edit', args=[controller.id]), {
@@ -188,12 +257,18 @@ class ViewsTestCase(ControllerBaseTestCase):
                 'env-INITIAL_FORMS': 0,
                 'env-MIN_NUM_FORMS': 0,
                 'env-MAX_NUM_FORMS': 100,
+                'label-TOTAL_FORMS': 0,
+                'label-INITIAL_FORMS': 0,
+                'label-MIN_NUM_FORMS': 0,
+                'label-MAX_NUM_FORMS': 100,
+                'webhook_token': webhook_token,
             })
         controller = Controller.objects.get(pk=controller.id)
         self.assertEqual(controller.marathon_cpus, 0.5)
         self.assertEqual(controller.marathon_mem, 100.0)
         self.assertEqual(controller.marathon_instances, 2)
         self.assertEqual(controller.marathon_cmd, '/path/to/exec some command')
+        self.assertEqual(str(controller.webhook_token), webhook_token)
 
     @responses.activate
     def test_advanced_page_marathon_error(self):
@@ -210,6 +285,10 @@ class ViewsTestCase(ControllerBaseTestCase):
             'env-INITIAL_FORMS': 0,
             'env-MIN_NUM_FORMS': 0,
             'env-MAX_NUM_FORMS': 100,
+            'label-TOTAL_FORMS': 0,
+            'label-INITIAL_FORMS': 0,
+            'label-MIN_NUM_FORMS': 0,
+            'label-MAX_NUM_FORMS': 100,
         }
 
         response = self.client.post(reverse('base:add'), data)
@@ -229,6 +308,10 @@ class ViewsTestCase(ControllerBaseTestCase):
                 'env-INITIAL_FORMS': 0,
                 'env-MIN_NUM_FORMS': 0,
                 'env-MAX_NUM_FORMS': 100,
+                'label-TOTAL_FORMS': 0,
+                'label-INITIAL_FORMS': 0,
+                'label-MIN_NUM_FORMS': 0,
+                'label-MAX_NUM_FORMS': 100,
             })
         controller = Controller.objects.get(pk=controller.id)
         self.assertEqual(controller.marathon_cpus, 0.5)
@@ -244,15 +327,35 @@ class ViewsTestCase(ControllerBaseTestCase):
         self.assertNotContains(resp, 'Start new base controller')
         self.assertNotContains(resp, 'edit')
 
-        self.client.login(username='testuser2', password='test')
+        # normal user with org who is not admin
+        user = User.objects.create_user('joe2', 'joe2@email.com', '1234')
+        org = Organization.objects.get(pk=1)
+        OrganizationUserRelation.objects.create(user=user, organization=org)
+
+        self.client.login(username='joe2', password='1234')
 
         self.client.post(
             reverse('user_settings'), {'settings_level': 'expert'})
         resp = self.client.get(reverse('home'))
-        self.assertContains(resp, 'Basic')
-        self.assertContains(resp, 'Docker')
-        self.assertContains(resp, 'Free Basics')
-        self.assertContains(resp, 'card-content new-site')
+        self.assertContains(resp, 'You do not have permission to create')
+
+    def test_normal_user_who_is_org_admin_can_create_sites(self):
+        resp = self.client.get(reverse('home'))
+        self.assertNotContains(resp, 'Start new base controller')
+        self.assertNotContains(resp, 'edit')
+
+        # normal user with org who is not admin
+        user = User.objects.create_user('joe2', 'joe2@email.com', '1234')
+        org = Organization.objects.get(pk=1)
+        OrganizationUserRelation.objects.create(
+            user=user, organization=org, is_admin=True)
+
+        self.client.login(username='joe2', password='1234')
+
+        self.client.post(
+            reverse('user_settings'), {'settings_level': 'expert'})
+        resp = self.client.get(reverse('home'))
+        self.assertNotContains(resp, 'You do not have permission to create')
 
     def test_staff_access_required(self):
         self.mk_controller(controller={'owner': User.objects.get(pk=2)})
@@ -299,8 +402,8 @@ class ViewsTestCase(ControllerBaseTestCase):
                 settings.LOGDRIVER_PATH,
                 ('worker-machine-1/worker-machine-id'
                  '/frameworks/the-framework-id/executors'
-                 '/%s.the-task-id/runs/latest/stdout?n=0' %
-                 controller.app_id)))
+                 '/%s.the-task-id/runs/latest/stdout?n=%s' %
+                 (controller.app_id, settings.LOGDRIVER_BACKLOG))))
         self.assertEqual(resp['X-Accel-Buffering'], 'no')
 
     @responses.activate
@@ -321,8 +424,8 @@ class ViewsTestCase(ControllerBaseTestCase):
                 settings.LOGDRIVER_PATH,
                 ('worker-machine-1/worker-machine-id'
                  '/frameworks/the-framework-id/executors'
-                 '/%s.the-task-id/runs/latest/stderr?n=0' %
-                 controller.app_id)))
+                 '/%s.the-task-id/runs/latest/stderr?n=%s' %
+                 (controller.app_id, settings.LOGDRIVER_BACKLOG))))
         self.assertEqual(resp['X-Accel-Buffering'], 'no')
 
     @responses.activate
@@ -454,3 +557,93 @@ class ViewsTestCase(ControllerBaseTestCase):
         resp = self.client.get(reverse('home'))
         self.assertContains(resp, 'Failed to delete')
         self.assertEquals(Controller.objects.all().count(), 1)
+
+    @responses.activate
+    def test_app_webhook_restart(self):
+        """
+        The restart webhook restarts the app and requires no authentication.
+        """
+        token = uuid.uuid4()
+        controller = self.mk_controller(controller={
+            'owner': User.objects.get(pk=2),
+            'state': 'done',
+            'webhook_token': token,
+        })
+        self.mock_restart_marathon_app(controller.app_id)
+
+        resp = Client().post(
+            reverse('base:webhook_restart', args=[controller.id, token]))
+        self.assertEqual(resp.status_code, 200)
+        self.assertEqual(json.loads(resp.content), {})
+        self.assertEqual(len(responses.calls), 1)
+
+    @responses.activate
+    def test_app_webhook_restart_wrong_token(self):
+        """
+        The restart webhook 404s if the token doesn't match.
+        """
+        token = uuid.uuid4()
+        controller = self.mk_controller(controller={
+            'owner': User.objects.get(pk=2),
+            'state': 'done',
+            'webhook_token': token,
+        })
+        self.mock_restart_marathon_app(controller.app_id)
+
+        resp = Client().post(
+            reverse('base:webhook_restart', args=[controller.id, 'abc']))
+        self.assertEqual(resp.status_code, 404)
+        self.assertEqual(len(responses.calls), 0)
+
+    @responses.activate
+    def test_app_webhook_restart_no_token(self):
+        """
+        The restart webhook 404s if the controller has no token set.
+        """
+        controller = self.mk_controller(controller={
+            'owner': User.objects.get(pk=2),
+            'state': 'done'})
+        self.mock_restart_marathon_app(controller.app_id)
+
+        resp = Client().post(
+            reverse('base:webhook_restart', args=[controller.id, 'abc']))
+        self.assertEqual(resp.status_code, 404)
+        self.assertEqual(len(responses.calls), 0)
+
+    @responses.activate
+    def test_app_webhook_restart_no_get(self):
+        """
+        The restart webhook does not accept GET requests.
+        """
+        token = uuid.uuid4()
+        controller = self.mk_controller(controller={
+            'owner': User.objects.get(pk=2),
+            'state': 'done',
+            'webhook_token': token,
+        })
+        self.mock_restart_marathon_app(controller.app_id)
+
+        resp = Client().get(
+            reverse('base:webhook_restart', args=[controller.id, token]))
+        self.assertEqual(resp.status_code, 405)
+        self.assertEqual(len(responses.calls), 0)
+
+    @responses.activate
+    def test_app_webhook_restart_error(self):
+        """
+        The restart webhook returns an error if the restart failed for some
+        reason.
+        """
+        token = uuid.uuid4()
+        controller = self.mk_controller(controller={
+            'owner': User.objects.get(pk=2),
+            'state': 'done',
+            'webhook_token': token,
+        })
+        self.mock_restart_marathon_app(controller.app_id, 404)
+
+        resp = Client().post(
+            reverse('base:webhook_restart', args=[controller.id, token]))
+        self.assertEqual(resp.status_code, 500)
+        self.assertEqual(
+            json.loads(resp.content), {'error': 'Restart failed.'})