+++ mc2/controllers/base/views.py
@@ -3,7 +3,8 @@ import os.path
 from django.conf import settings
 from django.shortcuts import get_object_or_404, redirect
 from django.http import (
-    HttpResponse, HttpResponseNotFound, HttpResponseBadRequest)
+    HttpResponse, HttpResponseNotFound, HttpResponseBadRequest,
+    HttpResponseNotAllowed, HttpResponseServerError)
 from django.contrib.auth.decorators import (
     login_required, user_passes_test)
 from django.utils.decorators import method_decorator
@@ -88,6 +89,7 @@ class ControllerCreateView(ControllerViewMixin, CreateView):
         form.controller_form.instance.save()
 
         form.env_formset.instance = form.controller_form.instance
+        form.label_formset.instance = form.controller_form.instance
 
         response = super(ControllerCreateView, self).form_valid(form)
         tasks.start_new_controller.delay(form.controller_form.instance.id)
@@ -191,3 +193,32 @@ class ControllerDeleteView(ControllerViewMixin, View):
             messages.error(self.request, msg)
             return HttpResponseBadRequest(msg)
         return redirect('home')
+
+
+class ControllerWebhookRestartView(View):
+    """
+    Unauthenticated webhook to restart an app.
+
+    Technically, this is a 'URL capability'.
+    """
+
+    @method_decorator(csrf_exempt)
+    def dispatch(self, *args, **kwargs):
+        return super(
+            ControllerWebhookRestartView, self).dispatch(*args, **kwargs)
+
+    def get(self, request, controller_pk, token):
+        return HttpResponseNotAllowed(['POST'])
+
+    def post(self, request, controller_pk, token):
+        controller = get_object_or_404(Controller, pk=controller_pk)
+        if str(controller.webhook_token) != token:
+            return HttpResponseNotFound()
+
+        try:
+            controller.marathon_restart_app()
+        except exceptions.MarathonApiException:
+            return HttpResponseServerError(
+                json.dumps({'error': 'Restart failed.'}),
+                content_type='application/json')
+        return HttpResponse(json.dumps({}), content_type='application/json')