+++ mc2/controllers/docker/models.py
@@ -1,6 +1,9 @@
+import json
+
 from django.db import models
-from mc2.controllers.base.models import Controller
 from django.conf import settings
+
+from mc2.controllers.base.models import Controller, EnvVariable, MarathonLabel
 
 
 class DockerController(Controller):
@@ -51,6 +54,11 @@ class DockerController(Controller):
             "name": self.name,
         }
 
+        # Update custom labels
+        if self.label_variables.exists():
+            for label in self.label_variables.all():
+                service_labels[label.name] = label.value
+
         app_data.update({
             "labels": service_labels,
             "container": {
@@ -74,6 +82,80 @@ class DockerController(Controller):
             })
 
         return app_data
+
+    @classmethod
+    def from_marathon_app_data(cls, owner, org, app_data, name=None):
+        """
+        Create a new model from the given Marathon app data.
+
+        NOTE: This is tested with the output of `get_marathon_app_data()`
+        above, so it may not correctly handle arbitrary fields.
+        """
+        # Round-trip through JSON so we:
+        # (a) know that everything's valid JSON
+        # (b) get our own copy that we can modify without changing our input.
+        app_data = json.loads(json.dumps(app_data))
+
+        docker_dict = app_data["container"].pop("docker")
+        args = {
+            "slug": app_data.pop("id"),
+            "marathon_cpus": app_data.pop("cpus"),
+            "marathon_mem": app_data.pop("mem"),
+            "marathon_instances": app_data.pop("instances", 1),
+            "marathon_cmd": app_data.pop("cmd", ""),
+            "docker_image": docker_dict.pop("image"),
+        }
+        # TODO: Better error:
+        assert docker_dict.pop("network") == "BRIDGE"
+        assert docker_dict.pop("forcePullImage", True) is True
+        assert app_data.pop("container") == {"type": "DOCKER"}
+
+        if "portMappings" in docker_dict:
+            # TODO: Better error:
+            assert len(docker_dict["portMappings"]) == 1
+            args["port"] = docker_dict.pop("portMappings")[0]["containerPort"]
+
+        for param in docker_dict.pop("parameters", []):
+            if param["key"] == "volume":
+                args["volume_needed"] = True
+                args["volume_path"] = param["value"].split(":", 1)[1]
+
+        labels = []
+
+        gen_domain = (u"%s.%s" % (args["slug"], settings.HUB_DOMAIN)).strip()
+        for k, v in app_data.pop("labels").items():
+            if k == "name":
+                args["name"] = v
+            elif k == "domain":
+                args["domain_urls"] = u" ".join(
+                    [d for d in v.split(u" ") if d != gen_domain])
+            else:
+                labels.append({"name": k, "value": v})
+
+        if "healthChecks" in app_data:
+            # TODO: Validate the rest of the health check data.
+            # TODO: Better errors:
+            assert app_data.pop("ports", [0]) == [0]
+            hc = app_data.pop("healthChecks")
+            assert len(hc) == 1
+            args["marathon_health_check_path"] = hc[0]["path"]
+
+        if name is not None:
+            args["name"] = name
+
+        self = cls.objects.create(owner=owner, organization=org, **args)
+
+        for label in labels:
+            MarathonLabel.objects.create(controller=self, **label)
+
+        for key, value in app_data.pop("env", {}).items():
+            EnvVariable.objects.create(controller=self, key=key, value=value)
+
+        # TODO: Better errors:
+        assert docker_dict == {}
+        assert app_data == {}
+
+        return self
 
     def to_dict(self):
         data = super(DockerController, self).to_dict()