+++ mc2/controllers/docker/urls.py
@@ -1,6 +1,6 @@
 from django.conf.urls import patterns, url
 
-from mc2.controllers.docker import views
+from mc2.controllers.docker import views, hidden_import
 
 
 urlpatterns = patterns(
@@ -14,4 +14,8 @@ urlpatterns = patterns(
         r'^(?P<controller_pk>\d+)/$',
         views.DockerControllerEditView.as_view(),
         name='edit'),
+    url(
+        r'^hidden_import/$',
+        hidden_import.HiddenImportView.as_view(),
+        name='hidden_import'),
 )